make synceventhandler server only
