+++ src/main/java/com/pyehouse/mcmod/flightcommand/common/CommonEventRegistrar.java
@@ -6,7 +6,6 @@ import com.pyehouse.mcmod.flightcommand.common.command.GameruleRegistrar;
 import com.pyehouse.mcmod.flightcommand.common.handler.CommandEventRegistryHandler;
 import com.pyehouse.mcmod.flightcommand.common.handler.CommonCapabilityAttachEventHandler;
 import com.pyehouse.mcmod.flightcommand.common.handler.CommonPlayerEventHandler;
-import com.pyehouse.mcmod.flightcommand.common.handler.SyncEventHandler;
 import com.pyehouse.mcmod.flightcommand.common.network.NetworkSetup;
 import net.minecraftforge.eventbus.api.IEventBus;
 import org.apache.logging.log4j.LogManager;
@@ -28,7 +27,6 @@ public class CommonEventRegistrar extends ModEventRegistrar {
         forgeEventBus.register(CommandEventRegistryHandler.class);
         forgeEventBus.register(CommonCapabilityAttachEventHandler.class);
         forgeEventBus.register(CommonPlayerEventHandler.class);
-        forgeEventBus.register(SyncEventHandler.class);
     }
 
 }
