+++ src/main/java/com/pyehouse/mcmod/flightcommand/server/ServerEventRegistrar.java
@@ -1,6 +1,7 @@
 package com.pyehouse.mcmod.flightcommand.server;
 
 import com.pyehouse.mcmod.flightcommand.api.util.ModEventRegistrar;
+import com.pyehouse.mcmod.flightcommand.server.handler.SyncEventHandler;
 import net.minecraftforge.eventbus.api.IEventBus;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -13,5 +14,6 @@ public class ServerEventRegistrar extends ModEventRegistrar {
     }
 
     public void registration() {
+        forgeEventBus.register(SyncEventHandler.class);
     }
 }
