+++ src/main/java/com/pyehouse/mcmod/flightcommand/server/handler/SyncEventHandler.java
@@ -1,5 +1,7 @@
-package com.pyehouse.mcmod.flightcommand.common.handler;
+package com.pyehouse.mcmod.flightcommand.server.handler;
 
+import com.pyehouse.mcmod.flightcommand.api.capability.FlightCapability;
+import com.pyehouse.mcmod.flightcommand.api.capability.IFlightCapability;
 import com.pyehouse.mcmod.flightcommand.common.network.ClientUpdater;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraftforge.event.entity.player.PlayerEvent;
@@ -28,5 +30,19 @@ public class SyncEventHandler {
     @SubscribeEvent
     public static void playerRespawn(PlayerEvent.PlayerRespawnEvent event) {
         syncPlayerData(event.getPlayer());
+    }
+
+    @SubscribeEvent
+    public static void onPlayerClone(PlayerEvent.Clone event) {
+        PlayerEntity player = event.getPlayer();
+        IFlightCapability flightCap = player.getCapability(FlightCapability.CAPABILITY_FLIGHT).orElse(null);
+        if (flightCap == null) {
+            LOGGER.error("Missing IFlightCapability on player");
+            return;
+        }
+        IFlightCapability oldFlightCap = event.getOriginal().getCapability(FlightCapability.CAPABILITY_FLIGHT).orElse(null);
+        if (oldFlightCap != null) {
+            flightCap.copyFrom(oldFlightCap);
+        }
     }
 }
